move optimization back to azure.quantum, organize targets by provider
