+++ azure-quantum/azure/quantum/optimization/problem.py
@@ -11,9 +11,9 @@ import json
 import numpy
 import os
 
-from typing import List, Tuple, Union, Dict, Optional, Type, TYPE_CHECKING
+from typing import List, Tuple, Union, Dict, Optional, TYPE_CHECKING
 from enum import Enum
-from azure.quantum.target.optimization import TermBase, Term, GroupType, GroupedTerm
+from azure.quantum.optimization import TermBase, Term, GroupType, GroupedTerm
 from azure.quantum.storage import (
     ContainerClient,
     download_blob,
