+++ azure-quantum/azure/quantum/optimization/term.py
