+++ azure-quantum/tests/unit/test_online_problem.py
@@ -1,14 +1,14 @@
 #!/bin/env python
 # -*- coding: utf-8 -*-
 ##
-# test_online_problem.py: Checks correctness of azure.quantum.target.optimization module.
+# test_online_problem.py: Checks correctness of azure.quantum.optimization module.
 ##
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 ##
 from unittest.mock import Mock, patch
-from azure.quantum.target.optimization import Problem, OnlineProblem
-import azure.quantum.target.optimization.problem
+from azure.quantum.optimization import Problem, OnlineProblem
+import azure.quantum.optimization.problem
 from common import expected_terms, QuantumTestBase
 import json
 
@@ -21,14 +21,14 @@ class TestOnlineProblemClass(QuantumTestBase):
 
     # TODO: instead of using mock, connect to a live service and record the responses
     def test_download(self):
-        with patch("azure.quantum.target.optimization.problem.download_blob") as mock_download_blob,\
-            patch("azure.quantum.target.optimization.problem.BlobClient") as mock_blob_client, \
-            patch("azure.quantum.target.optimization.problem.ContainerClient") as mock_container_client:
+        with patch("azure.quantum.optimization.problem.download_blob") as mock_download_blob,\
+            patch("azure.quantum.optimization.problem.BlobClient") as mock_blob_client, \
+            patch("azure.quantum.optimization.problem.ContainerClient") as mock_container_client:
             mock_download_blob.return_value=expected_terms()
             mock_blob_client.from_blob_url.return_value = Mock()
             mock_container_client.from_container_url.return_value = Mock()
             actual_result = self.o_problem.download(self.mock_ws)
             # TODO: add test that user warning was registered in log
             assert actual_result.name == "test"
-            azure.quantum.target.optimization.problem.download_blob.assert_called_once()
+            azure.quantum.optimization.problem.download_blob.assert_called_once()
             assert isinstance(actual_result, Problem)
