+++ azure-quantum/tests/unit/test_target.py
@@ -3,7 +3,7 @@ import warnings
 
 from azure.core.exceptions import HttpResponseError
 from azure.quantum.job.job import Job
-from azure.quantum.target import IonQ, Honeywell, MicrosoftFleetManagement
+from azure.quantum.target import IonQ, Honeywell, FleetManagement
 
 from common import QuantumTestBase, ZERO_UID
 
@@ -160,7 +160,7 @@ class TestHoneywell(QuantumTestBase):
                     assert results["c0"] == ["0"]
                     assert results["c1"] == ["000"]
 
-class TestMicrosoftFleetManagement(QuantumTestBase):
+class TestFleetManagement(QuantumTestBase):
     mock_create_job_id_name = "create_job_id"
     create_job_id = Job.create_job_id
 
@@ -187,7 +187,7 @@ class TestMicrosoftFleetManagement(QuantumTestBase):
                 }
             }
 
-            target = MicrosoftFleetManagement(workspace=workspace)
+            target = FleetManagement(workspace=workspace)
             try:
                 job = target.submit(input=input, input_params=input_params)
             except HttpResponseError as e:
@@ -201,7 +201,7 @@ class TestMicrosoftFleetManagement(QuantumTestBase):
                         # Set a timeout for Honeywell recording
                         job.wait_until_completed(timeout_secs=60)
                     except TimeoutError:
-                        warnings.warn("MicrosoftFleetManagement execution exceeded timeout. Skipping fetching results.")
+                        warnings.warn("FleetManagement execution exceeded timeout. Skipping fetching results.")
                     else:
                         # Check if job succeeded
                         self.assertEqual(True, job.has_completed())
